move configuration parsing to server. Require step to allow reloading configuration via CLI
