+++ cmd/meeseeks/utils_test.go
@@ -3,77 +3,8 @@ package main
 import (
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
-
-	"github.com/GustavoCaso/meeseeks/internal/config"
-	"github.com/GustavoCaso/meeseeks/internal/logger"
 )
-
-func TestCreateProgramFromConfig(t *testing.T) {
-	tests := []struct {
-		name         string
-		config       config.ProgramConfig
-		expectedName string
-		expectError  bool
-		errorMessage string
-	}{
-		{
-			name: "basic program",
-			config: config.ProgramConfig{
-				Name:    "test-basic",
-				Command: "echo",
-				Args:    []string{"hello"},
-			},
-			expectedName: "test-basic",
-		},
-		{
-			name: "program with environment variables",
-			config: config.ProgramConfig{
-				Name:    "test-env",
-				Command: "bash",
-				Args:    []string{"-c", "echo $TEST_VAR"},
-				Env:     []string{"TEST_VAR=test_value"},
-			},
-			expectedName: "test-env",
-		},
-		{
-			name: "program with interval",
-			config: config.ProgramConfig{
-				Name:     "test-interval",
-				Command:  "echo",
-				Interval: "30s",
-			},
-			expectedName: "test-interval",
-		},
-	}
-
-	logger := logger.New()
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			prog, err := createProgramFromConfig(tt.config, logger)
-
-			if tt.expectError {
-				if err == nil {
-					t.Fatalf("Expected error but got none")
-				}
-				if !strings.Contains(err.Error(), tt.errorMessage) {
-					t.Fatalf("Expected error containing %q, got %q", tt.errorMessage, err.Error())
-				}
-				return
-			}
-
-			if err != nil {
-				t.Fatalf("Unexpected error: %v", err)
-			}
-
-			if prog.Name() != tt.expectedName {
-				t.Fatalf("Expected name %q, got %q", tt.expectedName, prog.Name())
-			}
-		})
-	}
-}
 
 func TestGetMeeseeksDir(t *testing.T) {
 	t.Run("uses environment variable when set", func(t *testing.T) {
